Refactored to use a shared package
