+++ internal/component/cloudflare/worker-src/.gitignore
@@ -143,6 +143,7 @@ out
 
 # Nuxt.js build / generate output
 .nuxt
+dist
 
 # Gatsby files
 .cache/
